improve interactions with the hover ring
When the user mouses over the middle of a cell, we've been completely
hiding the hover ring. I've found this to be pretty distracting though,
especially since moving between the input to output sides of a cell
causes the hover ring to flicker briefly.

This change makes it so the hover ring is shown even when the user's
mouse is directly over a cell. To make it more obvious that there's
special functionality available if the user clicks on the hover ring
itself, we can use a faint stroke color when the mouse is over the cell
and a more prominent highlight color when the mouse is over the hover
ring.

In addition to using a more prominent color, we can also use the more
prominent color on only half of the hover ring based on which side of
the cell the user is hovering over. This makes it more obvious that the
side of the hover ring that a user clicks on matters.